[FOLD] rekey test and work
